web/requirements.txt
Django==2.2.11
djangorestframework==3.10.3
django-filter==2.2.0
Markdown
requests==2.23.0
plotly
gunicorn
gevent
pandas
psycopg2
